This is a new pull request following up with:
- https://github.com/coreos/etcd/pull/3575
- https://github.com/coreos/etcd/pull/3582

And the goal of this PR is, as @yichengq suggests, `use map to reduce cancel cost,
and benchmark its difference`. As you see in my [example benchmark](https://github.com/gyuho/learn/blob/master/doc/go_function_method_pointer_nil_map_slice/slice_vs_map/benchmark_results.txt) [1], theoretically Go map should be faster in
deletion. And current watchable store does deletion in its unsynced slice with
cancel function. So I expected map version to be faster, if not slower:

```
go test -bench='BenchmarkWatchableStoreUnsynced' -benchmem -cpu 1,2,4,8 > slice.txt;

go test -bench='BenchmarkWatchableStoreUnsynced' -benchmem -cpu 1,2,4,8 > map.txt;

go get -v -u golang.org/x/tools/cmd/benchcmp;
benchcmp slice.txt map.txt > benchmark_results.txt;
```

<br>
Results (old is original code with slice, **negative** means `map` version is **faster** than `slice` version):

```
benchmark                              old ns/op     new ns/op     delta
BenchmarkWatchableStoreUnsynced        15358884      8238657       -46.36%
BenchmarkWatchableStoreUnsynced-2      2672428       2701123       +1.07%
BenchmarkWatchableStoreUnsynced-4      2253534       2445470       +8.52%
BenchmarkWatchableStoreUnsynced-8      2139670       2103291       -1.70%

benchmark                              old allocs     new allocs     delta
BenchmarkWatchableStoreUnsynced        10912          10330          -5.33%
BenchmarkWatchableStoreUnsynced-2      9717           9843           +1.30%
BenchmarkWatchableStoreUnsynced-4      10273          10241          -0.31%
BenchmarkWatchableStoreUnsynced-8      9889           7906           -20.05%

benchmark                              old bytes     new bytes     delta
BenchmarkWatchableStoreUnsynced        404238        386040        -4.50%
BenchmarkWatchableStoreUnsynced-2      356543        366115        +2.68%
BenchmarkWatchableStoreUnsynced-4      373095        382577        +2.54%
BenchmarkWatchableStoreUnsynced-8      359913        299902        -16.67%

```

So, my question is:

1. Do you think my benchmark is valid to compare `slice` and `map` in
   watchable store?
2. If not, what would be the best way to test real-life cases for `unsynced`?

Thanks! And please let me know if you have any feedback.

---

1. https://github.com/gyuho/learn/blob/master/doc/go_function_method_pointer_nil_map_slice/slice_vs_map/benchmark_results.txt