@gyuho 2015-09-25
New benchmark results (old is `slice`, new is `map`)

TEST 1:

```
benchmark                             old ns/op     new ns/op     delta
BenchmarkWatchableStoreUnsynced       72181         82401         +14.16%
BenchmarkWatchableStoreUnsynced-2     103256        56468         -45.31%
BenchmarkWatchableStoreUnsynced-4     54129         61255         +13.16%
BenchmarkWatchableStoreUnsynced-8     61312         52028         -15.14%

benchmark                             old allocs     new allocs     delta
BenchmarkWatchableStoreUnsynced       60             56             -6.67%
BenchmarkWatchableStoreUnsynced-2     61             55             -9.84%
BenchmarkWatchableStoreUnsynced-4     55             54             -1.82%
BenchmarkWatchableStoreUnsynced-8     53             55             +3.77%

benchmark                             old bytes     new bytes     delta
BenchmarkWatchableStoreUnsynced       2584          2470          -4.41%
BenchmarkWatchableStoreUnsynced-2     2597          2444          -5.89%
BenchmarkWatchableStoreUnsynced-4     2425          2403          -0.91%
BenchmarkWatchableStoreUnsynced-8     2363          2403          +1.69%
```

<br>
TEST 2:

```
benchmark                             old ns/op     new ns/op     delta
BenchmarkWatchableStoreUnsynced       88385         75425         -14.66%
BenchmarkWatchableStoreUnsynced-2     64586         62194         -3.70%
BenchmarkWatchableStoreUnsynced-4     55818         55031         -1.41%
BenchmarkWatchableStoreUnsynced-8     60633         54193         -10.62%

benchmark                             old allocs     new allocs     delta
BenchmarkWatchableStoreUnsynced       54             54             +0.00%
BenchmarkWatchableStoreUnsynced-2     54             56             +3.70%
BenchmarkWatchableStoreUnsynced-4     56             55             -1.79%
BenchmarkWatchableStoreUnsynced-8     54             55             +1.85%

benchmark                             old bytes     new bytes     delta
BenchmarkWatchableStoreUnsynced       2370          2385          +0.63%
BenchmarkWatchableStoreUnsynced-2     2388          2471          +3.48%
BenchmarkWatchableStoreUnsynced-4     2480          2403          -3.10%
BenchmarkWatchableStoreUnsynced-8     2395          2407          +0.50%
```


Thanks @yichengq.

Please let me know if you have any other feedback!