@gyuho 2015-09-25
I ran the same test in Go 1.5

- Go version: `go version go1.5.1 linux/amd64`
- Machine: Intel(R) Core(TM) i7-4750HQ CPU @ 2.00GHz


TEST 1:

```
benchmark                             old ns/op     new ns/op     delta
BenchmarkWatchableStoreUnsynced       87295         68613         -21.40%
BenchmarkWatchableStoreUnsynced-2     57660         54475         -5.52%
BenchmarkWatchableStoreUnsynced-4     58272         56447         -3.13%
BenchmarkWatchableStoreUnsynced-8     43593         52191         +19.72%

benchmark                             old allocs     new allocs     delta
BenchmarkWatchableStoreUnsynced       50             51             +2.00%
BenchmarkWatchableStoreUnsynced-2     51             51             +0.00%
BenchmarkWatchableStoreUnsynced-4     51             52             +1.96%
BenchmarkWatchableStoreUnsynced-8     51             51             +0.00%

benchmark                             old bytes     new bytes     delta
BenchmarkWatchableStoreUnsynced       2556          2585          +1.13%
BenchmarkWatchableStoreUnsynced-2     2573          2563          -0.39%
BenchmarkWatchableStoreUnsynced-4     2571          2637          +2.57%
BenchmarkWatchableStoreUnsynced-8     2620          2615          -0.19%
```

<br>
TEST 2:

```
benchmark                             old ns/op     new ns/op     delta
BenchmarkWatchableStoreUnsynced       78532         91794         +16.89%
BenchmarkWatchableStoreUnsynced-2     62161         59757         -3.87%
BenchmarkWatchableStoreUnsynced-4     52915         59778         +12.97%
BenchmarkWatchableStoreUnsynced-8     53171         53726         +1.04%

benchmark                             old allocs     new allocs     delta
BenchmarkWatchableStoreUnsynced       49             50             +2.04%
BenchmarkWatchableStoreUnsynced-2     49             52             +6.12%
BenchmarkWatchableStoreUnsynced-4     52             53             +1.92%
BenchmarkWatchableStoreUnsynced-8     52             51             -1.92%

benchmark                             old bytes     new bytes     delta
BenchmarkWatchableStoreUnsynced       2515          2568          +2.11%
BenchmarkWatchableStoreUnsynced-2     2529          2648          +4.71%
BenchmarkWatchableStoreUnsynced-4     2664          2659          -0.19%
BenchmarkWatchableStoreUnsynced-8     2630          2600          -1.14%

```