@yichengq 2015-09-25
I asked the question `Is this expected?` because the cancel var is assigned 301 times, but is only used once.

I would expect map performs well than slice in the case that cancel one watcher when there are thousands, maybe millions of watchers. But I don't know whether there is real case for it.

For your first benchmark, it actually benchmarks the operation that 301 put + 1 cancel. For the second one, it benchmarks 1 put + 1 cancel when async watcher list is empty, which is expected to have similar performance.