@gyuho 2015-09-25
@yichengq You are right. 

Sorry, I am still trying to understand how the watcher works
and trying to find the best way to evaluate this change with benchmarks.

So back to the original todo: `TODO: use map to reduce cancel cost`.
And `cancel` needs to find and delete the global reference of watcher
in the pool of `unsynced` watchers (from either `slice` or `map`).

So regarding `cancel`, we need faster deletion
(https://github.com/coreos/etcd/blob/master/storage/watchable_store.go#L178):
With millions of watchers in the `unsynced`, cancel with `map`
would be faster. And anything else should perform about the same
(append to slice vs. update/set map).

So, in the first Benchmark, I just wanted to populate as many watchers
as I can, and call the `cancel` function, to see how it performs.

```go
func BenchmarkWatchableStoreUnsynced(b *testing.B) {
	s := newWatchableStore(tmpPath)
	defer cleanup(s, tmpPath)

	for i := 0; i < 300; i++ {
		s.Put([]byte(fmt.Sprint("foo", i)), []byte("bar"))
	}

	b.ReportAllocs()
	b.StartTimer()

	for i := 0; i < b.N; i++ {
		_, cancel := s.Watcher([]byte(fmt.Sprint("foo", i)), true, -1, 0)
		for i := 0; i < 300; i++ {
			_, cancel = s.Watcher([]byte(fmt.Sprint("foo", i)), true, -1, 0)
		}
		cancel()
	}
}

```

But think I was doing it wrong, as you pointed out. I am calling `cancel` only once
for every 300 puts.

So I changed the test to cancel for every watcher.
I put 1042 keys and 1042 watchers. 1042 is just an arbitrary number,
and small enough to get reasonable number of iteration from benchmarks.

And here's the test results from the new benchmarks. 

TEST 1:

```
benchmark                                             old ns/op     new ns/op     delta
BenchmarkWatchableStoreUnsyncedCancel                 70518397      63562826      -9.86%
BenchmarkWatchableStoreUnsyncedCancel-2               19837178      14149584      -28.67%
BenchmarkWatchableStoreUnsyncedCancel-4               14587545      12747152      -12.62%
BenchmarkWatchableStoreUnsyncedCancel-8               16456413      13681624      -16.86%
BenchmarkWatchableStoreUnsyncedCancelSequential       15462569      14382688      -6.98%
BenchmarkWatchableStoreUnsyncedCancelSequential-2     12652903      13629045      +7.71%
BenchmarkWatchableStoreUnsyncedCancelSequential-4     11968235      12606191      +5.33%
BenchmarkWatchableStoreUnsyncedCancelSequential-8     13254876      11395397      -14.03%

benchmark                                             old allocs     new allocs     delta
BenchmarkWatchableStoreUnsyncedCancel                 58109          58888          +1.34%
BenchmarkWatchableStoreUnsyncedCancel-2               46496          42800          -7.95%
BenchmarkWatchableStoreUnsyncedCancel-4               42907          42691          -0.50%
BenchmarkWatchableStoreUnsyncedCancel-8               44679          42611          -4.63%
BenchmarkWatchableStoreUnsyncedCancelSequential       41367          41640          +0.66%
BenchmarkWatchableStoreUnsyncedCancelSequential-2     41574          43283          +4.11%
BenchmarkWatchableStoreUnsyncedCancelSequential-4     41863          42511          +1.55%
BenchmarkWatchableStoreUnsyncedCancelSequential-8     42609          41623          -2.31%

benchmark                                             old bytes     new bytes     delta
BenchmarkWatchableStoreUnsyncedCancel                 14179119      13027342      -8.12%
BenchmarkWatchableStoreUnsyncedCancel-2               5454540       3206899       -41.21%
BenchmarkWatchableStoreUnsyncedCancel-4               3026311       2654940       -12.27%
BenchmarkWatchableStoreUnsyncedCancel-8               4219613       2843076       -32.62%
BenchmarkWatchableStoreUnsyncedCancelSequential       2390279       2506617       +4.87%
BenchmarkWatchableStoreUnsyncedCancelSequential-2     2470202       3571160       +44.57%
BenchmarkWatchableStoreUnsyncedCancelSequential-4     2661278       2901270       +9.02%
BenchmarkWatchableStoreUnsyncedCancelSequential-8     3126862       2410299       -22.92%

```

<br>

TEST 2:

```
benchmark                                             old ns/op     new ns/op     delta
BenchmarkWatchableStoreUnsyncedCancel                 61413914      64065904      +4.32%
BenchmarkWatchableStoreUnsyncedCancel-2               21181668      17743837      -16.23%
BenchmarkWatchableStoreUnsyncedCancel-4               14419478      12279642      -14.84%
BenchmarkWatchableStoreUnsyncedCancel-8               13349545      16010846      +19.94%
BenchmarkWatchableStoreUnsyncedCancelSequential       15243314      15797653      +3.64%
BenchmarkWatchableStoreUnsyncedCancelSequential-2     12592730      12963295      +2.94%
BenchmarkWatchableStoreUnsyncedCancelSequential-4     12119024      11609582      -4.20%
BenchmarkWatchableStoreUnsyncedCancelSequential-8     11587317      11636739      +0.43%

benchmark                                             old allocs     new allocs     delta
BenchmarkWatchableStoreUnsyncedCancel                 58208          57134          -1.85%
BenchmarkWatchableStoreUnsyncedCancel-2               46460          45976          -1.04%
BenchmarkWatchableStoreUnsyncedCancel-4               43268          42483          -1.81%
BenchmarkWatchableStoreUnsyncedCancel-8               42626          44342          +4.03%
BenchmarkWatchableStoreUnsyncedCancelSequential       41554          41866          +0.75%
BenchmarkWatchableStoreUnsyncedCancelSequential-2     41708          42489          +1.87%
BenchmarkWatchableStoreUnsyncedCancelSequential-4     41829          41598          -0.55%
BenchmarkWatchableStoreUnsyncedCancelSequential-8     41620          41535          -0.20%

benchmark                                             old bytes     new bytes     delta
BenchmarkWatchableStoreUnsyncedCancel                 12725402      13066011      +2.68%
BenchmarkWatchableStoreUnsyncedCancel-2               6427817       4989710       -22.37%
BenchmarkWatchableStoreUnsyncedCancel-4               3381741       2503254       -25.98%
BenchmarkWatchableStoreUnsyncedCancel-8               2709035       4019096       +48.36%
BenchmarkWatchableStoreUnsyncedCancelSequential       2430769       2778142       +14.29%
BenchmarkWatchableStoreUnsyncedCancelSequential-2     2565613       2944004       +14.75%
BenchmarkWatchableStoreUnsyncedCancelSequential-4     2608552       2419854       -7.23%
BenchmarkWatchableStoreUnsyncedCancelSequential-8     2412501       2409059       -0.14%

```


But I am still not sure what would be the real use case and not sure about
the best way to benchmark. Test is still a bit noisy.

Can you help me write a better test for this? 

Thanks for the review again.