@gyuho 2015-10-20
/cc @yichengq I updated the benchmark tests. Please review.


storage/watchable_store: use map for unsynced

This is for `use map to reduce cancel cost, and benchmark its difference`.
Please see below for benchmark results ([1], [2]). `map` implementation for
unsynced performs better.



<br>
<br>
<hr>

[1]:

```
benchmark                                   old ns/op     new ns/op     delta
BenchmarkWatchableStoreUnsyncedCancel       21069         24129         +14.52%
BenchmarkWatchableStoreUnsyncedCancel-2     9753          3467          -64.45%
BenchmarkWatchableStoreUnsyncedCancel-4     30084         2270          -92.45%
BenchmarkWatchableStoreUnsyncedCancel-8     18404         4407          -76.05%

benchmark                                   old allocs     new allocs     delta
BenchmarkWatchableStoreUnsyncedCancel       31             31             +0.00%
BenchmarkWatchableStoreUnsyncedCancel-2     14             6              -57.14%
BenchmarkWatchableStoreUnsyncedCancel-4     8              3              -62.50%
BenchmarkWatchableStoreUnsyncedCancel-8     9              2              -77.78%

benchmark                                   old bytes     new bytes     delta
BenchmarkWatchableStoreUnsyncedCancel       1442          1522          +5.55%
BenchmarkWatchableStoreUnsyncedCancel-2     630           431           -31.59%
BenchmarkWatchableStoreUnsyncedCancel-4     332           136           -59.04%
BenchmarkWatchableStoreUnsyncedCancel-8     369           211           -42.82%
```



[2]:

```
benchmark                                   old ns/op     new ns/op     delta
BenchmarkWatchableStoreUnsyncedCancel       23910         24476         +2.37%
BenchmarkWatchableStoreUnsyncedCancel-2     10119         1747          -82.74%
BenchmarkWatchableStoreUnsyncedCancel-4     18164         1613          -91.12%
BenchmarkWatchableStoreUnsyncedCancel-8     11739         3497          -70.21%

benchmark                                   old allocs     new allocs     delta
BenchmarkWatchableStoreUnsyncedCancel       32             31             -3.12%
BenchmarkWatchableStoreUnsyncedCancel-2     16             2              -87.50%
BenchmarkWatchableStoreUnsyncedCancel-4     9              1              -88.89%
BenchmarkWatchableStoreUnsyncedCancel-8     15             6              -60.00%

benchmark                                   old bytes     new bytes     delta
BenchmarkWatchableStoreUnsyncedCancel       1499          1547          +3.20%
BenchmarkWatchableStoreUnsyncedCancel-2     593           280           -52.78%
BenchmarkWatchableStoreUnsyncedCancel-4     369           72            -80.49%
BenchmarkWatchableStoreUnsyncedCancel-8     580           241           -58.45%
```